pyproject.toml
[project]
name = "trees"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jax>=0.6.0",
    "polars>=1.26.0",
    "streamlit>=1.42.2",
    "streamlit-flow-component>=1.6.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.6",
    "ipython",
]

[tool.ruff]
line-length = 100
ignore = ["COM812"]
